[ALPS07147215] mt6879: audio: fix coverity
1. Add range check to avoid passing negative value as array index
2. Give variable initial value
3. Component in probe function never null, dont need to check

MTK-Commit-Id: 352a0d3fa3da60a86517101d0dab9edccd3917d6

CR-Id: ALPS07147215
Feature: alsa_sound_soc

+++ drivers/misc/mediatek/audio_ipi/common/audio_ipi_platform_common.c
@@ -255,11 +255,11 @@ uint32_t audio_get_audio_ipi_id_by_dsp(const uint32_t dsp_id)
 
 	if (is_audio_use_adsp(dsp_id)) {
 #if IS_ENABLED(CONFIG_MTK_AUDIODSP_SUPPORT)
-		audio_ipi_id = ADSP_IPI_AUDIO;
+		audio_ipi_id = (uint32_t) ADSP_IPI_AUDIO;
 #endif
 	} else if (is_audio_use_scp(dsp_id)) {
 #if IS_ENABLED(CONFIG_MTK_SCP_AUDIO)
-		audio_ipi_id = SCP_AUDIO_IPI_AUDIO;
+		audio_ipi_id = (uint32_t) SCP_AUDIO_IPI_AUDIO;
 #endif
 	} else
 		pr_notice("dsp_id %u not support!!", dsp_id);
